requirements.txt
numpy
scipy
tensorflow-gpu>=1.12
nose2
